Updates Twitter shortlink regex
The previously used regex `/http\S*/` was a tad too liberal. It causes any punctuation that immediately follows the link to be included in the match rendering the link invalid. Switching to `https*:\/\/t.co\/\w+` seems to remedy this, but I may need to revisit this again as I'd like to display the actual link (and/or the linked content) in the future.